updated RatingViewModel to follow the MVVM pattern correctly
