+++ app/src/main/java/com/example/ventura/view/MapActivity.kt
@@ -50,6 +50,7 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import org.json.JSONObject
 import android.app.AlertDialog
+import com.example.ventura.viewmodel.RatingViewModelFactory
 import com.example.ventura.viewmodel.UserPreferencesViewModelFactory
 
 class MapsActivity : AppCompatActivity() {
@@ -92,7 +93,7 @@ class MapsActivity : AppCompatActivity() {
 
         jsonViewModel = ViewModelProvider(this, CampusLocationsViewModelFactory(this)).get(CampusLocationsViewModel::class.java)
         userPrefViewModel = ViewModelProvider(this, UserPreferencesViewModelFactory(this)).get(UserPreferencesViewModel::class.java)
-        ratingViewModel = ViewModelProvider(this).get(RatingViewModel::class.java)
+        ratingViewModel = ViewModelProvider(this, RatingViewModelFactory()).get(RatingViewModel::class.java)
 
         // allows the user to swipe down to update locations
         swipeRefreshLayout = findViewById<SwipeRefreshLayout>(R.id.swipe_refresh_layout)
